move postFormatFunctions.ts to post/ui from shared/util
