+++ apps/client/src/feature/post/ui/PostDetailView.tsx
@@ -20,7 +20,7 @@ import { useGuardContext } from '@/shared/context/hooks';
 import { ServiceContext } from '@/shared/context/ServiceContext';
 import { TokenContext } from '@/shared/context/TokenContext';
 import { useRouteNavigation } from '@/shared/route/useRouteNavigation';
-import { getEmploymentStatus } from '@/util/postFormatFunctions';
+import { getEmploymentStatus } from '@/feature/post/ui/postFormatFunctions.ts';
 
 export const PostDetailView = ({ postId }: { postId: string }) => {
   const { postDetailData } = useGetPostDetail({ postId: postId });
