+++ apps/client/src/feature/post/ui/postFormatFunctions.ts
