+++ apps/client/src/feature/resume/ui/ResumeList.tsx
@@ -5,7 +5,7 @@ import { useGuardContext } from '@/shared/context/hooks';
 import { ServiceContext } from '@/shared/context/ServiceContext';
 import { TokenContext } from '@/shared/context/TokenContext';
 import { useRouteNavigation } from '@/shared/route/useRouteNavigation';
-import { getFormatDate } from '@/util/postFormatFunctions';
+import { getFormatDate } from '@/feature/post/ui/postFormatFunctions.ts';
 
 export const ResumeListView = () => {
   const { resumeListData } = useGetResumeList();
